requirements.txt
Django>=5.1,<6
django-ninja>=1.1.0
django-environ>=0.11.2
django-cors-headers>=4.4.0
psycopg[binary]>=3.2
Pillow>=10.0
django-storages[boto3]>=1.14.3
boto3>=1.34

# Rich text / Markdown (admin) + HTML import cleanup
bleach>=6.1.0
markdownify>=0.12.1
pillow-avif-plugin>=1.4.6
